Add test cases for FilterCommand
Add test cases for automated testing of FilterCommand.

The data used is read from the default data given in AB3. However, as tags are supposed to only be module names, might need to modify next time in order to have a more comprehensive/accurate test case.

For now, we will leave the data as it is, and use the data from AB3 to do testing for FilterCommand.