Change address to GitHub username field
There is a need to add a new field for GitHub
username in evolving AB3 to HackNet. We also
no longer need the address field in HackNet.

Since these two fields are similar, let's change
the address field to a github username field instead
of adding a separate field entirely.

Fixes: AY2122S2-CS2103T-W13-3/tp#22