Update edit command to use updated person's fields
When new fields are added into person class, they also need to be manually added and updated to editCommand and editCommandParser.

Update with changes made from PR #16. Allows the fields to be modified and updated in one place without having to update edit command and edit command parser.

The decoupling will help make things smoother in the future and the process more efficient. Chances of bugs are also reduced since everything will be in a single place. Reduces the need to make multiple changes when new fields are added.