Modify same field violation in addressBook to use isSamePerson instead
There's redundant codes for checking whether the person's edit data is already in the addressBook.

Modify to use the function isSamePerson instead and isSamePerson would now check the fields that should not be duplicated in the addressBook instead of just checking the name since people can have the same name.