Revert back to before the refactor
This project is awful. I can't do anything about it short of deleting it
and rewriting it from the ground up. So I'm just going to revert it to
the way it was and we'll have to deal with it.

I know this git commit does not follow the git convention. I don't care.
I hate it.