Add Name class, Show class, and Tag class
Modify the original address book class to fit Trackermon.

1. Create Name class to represent name of show.
2. Create Show class to represent shows within Trackermon
3. Create Tag class to represent tags of the show.

This provides a layer of abstraction allowing us to abstract away lower
level details. For example, the toString method of Show case would
return the string representation of the Show class.

Going to add a status component into the Show class in the future
iteration.