Add AddWindow UI
UI for adding a person is implemented in this commit.

Functionality is minimal -- you can type text into each of the boxes, but submitting does nothing for now.
Pressing cancel will hide the window and reset all fields, so this works as intended.

The next step would be to properly handle a submit request. Right now, it just hides the window.