Add Assigned Lesson List to PersonInfoPanel and refactor some methods In PersonInfoPanel, assigned lessons are not shown.
Assigned Lessons can now be listed in the PersonInfoPanel, although assignedLessons are not stored in Person yet.

Refactor some methods and add two new abstract classes to allow the ListPanel to show both Person and Lessons.
Create new methods to easily populate the ListPanel with either Person/Lessons. Same for the InfoPanel although LessonInfoPanel is not implemented yet.

Issues:
- Person does not store assigned classes so temporary placeholders are used instead