Add base classes for Log and UniqueLogList
Added classes and their unit tests for Log and UniqueLogList classes for
use in the model component to support logs functionality.

The key design point are:
1. Logs
	* Logs are considered identical if they have the same title.
	* Logs are considered equal if they have the same title AND
	  description.
	* The default description if no description is provided is an
	  empty string.

2. UniqueLogList
The unique log list is intended to be pointed to by a Person object,
such that each person has a list of logs. In this commit, basic list
functionalities of add, set, remove, contains have been added.

Future modifications to these two base classes will be for command-specific
functionalities.