Modify interface and implementation class for UserPrefs and AddressBook
The old interfaces and classes are preserved to prevent the need of
further changes within other components. Hence, some parts of Logic,
Storage as well as tests are modified to use the old interfaces
appropriately.

The old interfaces and classes should be deleted once all components are
not referencing it anymore.