(Untested) Implement classes for addlog feature
This is a big commit, plenty of changes, with the main aim to implement
all classes necessary for the addlog command. Notable classes
written/changes are:

* `AddLogCommand` class
* `AddLogCommand.AddLogDescriptor` class
* `AddLogCommandParser` class
* (Incomplete) unit and integration tests and test cases for the above three classes

The inclusion of the three classes integrates nicely into the existing
`Command` interface framework. A notable design choice was to use a
static nested `AddLogCommand.AddLogDescriptor` class, and the rationale
was to follow the analogous design in the `EditCommand` class, since it
was easily extendible and helped with cleaner encapsulation.

Note that test cases have not been fully completed and debugged.