(Untested) Implement deletelog feature
In this commit, support for `deletelog` feature has been added. Main
changes are described here:
1. implementation of `DeleteLogCommand` class.
2. implementation of `DeleteLogCommandParser` class.
3. appropriate extensions to `ArgumentMultiMap` and `AddressBookParser`
   classes to support new feature.

Main design choice to take note is the use of a nested
`DeleteLogDescriptor` in `DeleteLogCommand`, much like that in
`EditCommand` and `AddLogCommand`. The idea is to encapsulate details of
the deletion into the helper class so that encapsulation is cleaner and
SLAP is carried out as far as is reasonable.

Note that this commit is untested.