Add summarise feature
Summarise command helps user to have a brief overview of the data they
have. This is helpful to see how many covid cases there are in each
faculty instead of navigating through the entire list.

From the list of students, they are separated by their faculty and the
application will count how many students there are, how many students
with a certain covid status. This is shown to the user.

It is done this way because the list command cannot show a tabulated
data of the student body.