Remove unecessary field parsers
There are parsers for every field in Person, which comprehensive tests
for each of the parser. However, these parsers are are duplicates of the
same functionality as the constructors of the fields themselves, with
the only exception being a String.trim() call.

Removing these duplicated code improves the code quality and vastly
reduces the number of tests necessary.

As a step towards better code quality, let's remove all of these
duplicated, useless, worthless, trash, shit code. Doing so will improve
our efficiency, decrease our workload, and improve our sanity.