Fix bug that causes assigned task to not be stored.
In the current implementation, the assigned task will not be
stored after exit.

How to replicate:
Launch TAPA -> Use an assign command on a student -> Exit TAPA.

Output:
On relaunching TAPA, the assigned task will disappear as it is not
being stored in the database.

Problem:
JsonAdaptedPerson.java does not take into account the new TaskList
field of a Person, which causes the current .json storage file to
not include the TaskList.