Add sort command to sort across all candidate fields
Currently, the default list is loaded from the json file.
There is no functionality to reorder this list.

User is not able to organise the list in a specific numerical
or alphabetical order.

Let's create a new sort command, which parses in the field
that the user wants to sort the candidate list by. The list
will order candidates based on the in-built java comparator
on the particular field.