Add test cases to SortCommandParserTest
There are currently no test cases for the sort command.

This causes code lines relevant to the sort command implementation
to be left unchecked, and thus the code coverage is less than
the required threshold.

Let's add a SortCommandParserTest class with test cases to
check for invalid fields in the user input, empty user input,
as well as 'successful' test cases.