Modify delete command
Currently the `delete` command uses index to delete a person.

Due to the use of index, `list` is utilized to get the correct index first.
This results in an additional command being used making it slow to type out both commands.

With this modification, Hustlebook deletes a person using their name. This can be achieved with only one command now.