Modify edit command
Currently the `edit` command requests users for index to select and edit a person.

Due to the use of index, `list` is utilized to get the correct index first.
This results in the user doing two commands to fulfil one., making it time consuming.

With this modification, HustleBook edits a person using their name.