Modify edit info parser
No way to clear info of client through edit to show the default
value of "No further info"

Let's make changes to the info parser to allow user to key edit NAME
i/ where parameter is empty to clear the info - similar to clearing
tags.

Note: This change still should not allow i/ to be empty when adding
new person.