Create new attributes for Candidates
Things that have been added in this commit
* Create ApplicationStatus for the Candidate's application
* Create InterviewStatus for interview status

The work flow for adding a new candidate into the system remains the same. The user will only be able to edit the candidate's statuses after it is added into the system.

Next off, I will push a new commit for the new test cases.