Change priceRange fields to non-final, because priceRange can change, for  example when bargaining.
Add PropertyToBuy and PropertyToSell classes. These are contracts
stating the House the buyer or seller is buying or selling
respectively, and the priceRange they are willing to buy or sell it for.

Rename Package name from house to property as house is is within the
scope of property.

Add methods to PriceRange to compare prices.

Fix test cases by adding overriding equals methods.

Fix bug in equals method in Location.