Update attendance command
Update user guide with attendance command.
Update AttendanceCommand.java createAttendanceForPet() method.
Update JsonAddressBookStorage.java

Add a pet to address book will create a JSON file and a txt file for the pet if it has
not been created already.
Attendance command takes in an index, date, pick up time and drop off time.
Command will be parsed to LocalDate, LocalTime and LocalTime respectively.
createAttendanceForPet() command takes in the parsed commands, adds them as
test in pet's txt file, sorts the txt file, reads it and outputs it as JSON file.

To be implemented: Optional fields such as pu/ and do/, time command to display
the pickup and drop off time, how to display them in GUI.