Update UG and DG with sort command
Currently, user stories, use cases and features do not contain
information and details pertaining to the sort command.

The information is missing for users and developers, who are unable
to attain consolidated information on the sort command
feature/functionality.

Let's update the UG and DG to include the sort command in the
following sections:
* Use cases (DeveloperGuide)
* User stories (DeveloperGuide)
* Features (UserGuide)
* Command summary (UserGuide)

Additional note: this commit also standardises the term
`applicant(s)` to `candidate(s)` in the UserGuide and DeveloperGuide.