Add block attribute into Person class.
In order to identify an entry's block, this requires the addition of the block attribute. This is done through the introduction of the block class, Block.java.

To accommodate this change, logic classes (AddCommand.java, EditCommand.java) and test case files were edited.

There are no PRs dependent on this as this PR is intended for v1.3.

This PR is not the finalised graphical UI of the block attribute in the PersonCard. This would be addressed in subsequent PRs.