Press ENTER to submit
At the last `TextField`, pressing `ENTER` key submits the fields in AddWindow.

AddWindow's functionality (for adding more fields) will be extended further, to do more commands.

Some of the commands that will be implemented:
- addmodule
- status

These will allow for users to input the modules of the contact that they are adding, and their status.