Extend functionality of AddWindow
It is possible to add modules, and a person's corresponding status.

The logic is not finalised yet and there may still be some changes.

With this current implementation, a `Person` is added even if the optional fields result in an error (e.g. `status` not being `blacklist` or `favourite`, or a `Module` being of the wrong format).

A new idea to improve on this would be to prevent anything from being executed if it results in an error; so user has to fix all errors before `AddWindow` executes and closes.

I will probably change to the above soon.