Update Find command to handle fragmented show names
Updated find command to handle fragmented words because the user might key in an incomplete word in the command line but still expect a result to be shown.

Creating a new method in StringUtil.java allows us to handle these fragmented words. For example, containsFragmentedWordIgnoreCase("ABc def", "AB") == true.

It is done this way so as to handle fragmented words being able to be found in the category to be searched in.