Update Find command to handle status and tags
Updated find command to be able to search for status and tags. In addition, updated find to include searching via prefixes. E.g., find [n/NAME] [s/STATUS] [t/TAGS].

Extending the current classes like FindCommandParser and creating a new class like ShowContainsKeywordsPredicate allow us to implement a better search function as mentioned above.

It is done this way so as to allow the user to have an easier time to find the show name/status/tags he/she desires.