Update Help Window
Updated the Help Window to show a command summary of all the commands in Trackermon with its syntax because the user might forget what are the commands and wants a quick look at the commands without going into the user guide.

Creating a CommandBuilder class in ui package, a Stackpane in SceneBuilder, and updating HelpWindow.java allows us to create a TableView of the command summary and display to the user.

It is done this way so as to give the user quick access to the commands available on top of having the user guide.