Update FindCommandParser
Updated FindCommandParser.java to be able to handle precise find because the user might want to find the exact show or tag within the showlist in Trackermon.

Implementing a for loop within the name and tag predicate check allows the AND operation to be utilized when searching the keywords input by the user.

It is done this way so as to give the user more flexibility to find the exact show he/she wants in the show list.