Check Validity of PriceRange in a better way.
Previously, PriceRange's validity was checked when it is instantiated.
That is not desirable as we need to assert many exceptions when creating
test cases involving PriceRange.

Now, PriceRange checks validity similar to the checks on Client Phone.
That is, we use checkArgument and static isValidPriceRange separately,
 which is more effective.

Remove the setUpper, setLower, and updatePriceRange methods from
PriceRange.

This is because these edits should be done by the editCommand,
which will create a new Client with a new PriceRange when we edit the
PriceRange, and will check its validity again via the above-mentioned
measures.