Add NullPropertyToBuy
Before, we didn't have a way to represent the lack of a property to buy
without using null, which is not recommended.

Now, there is a null property class to act as a placeholder when the
user has just added a buyer, and also for testing purposes. This class
essentially has the same things as a regular property, but has different
equality in the sense that as long as 2 properties are NullProperties
they are equal. Also, I have changed some of the test classes to remove
the IllegalValueException and try catch.