Update schedule command to check for candidate availability
Currently, the schedule command does not check whether the user's
input for interview datetime matches the candidate's availability.

This allows interviews to be set for candidates on days that
the candidates are not available. Instead, this should be invalid
and should throw an error.

Let's implement checks to ensure the interview datetime matches the
candidate's availabilities. A new method through the
ModelManager and InterviewSchedule class will return true if the
interview datetime matches the candidate's availability, and false
otherwise. The schedule command will throw an error message if
there is no match.