Modify list command
Currently, list command has options to filter out the type of tasks
which is better suited for the find command.

The options are currently optional and hence, there no need for the
parser to check for the options.

Lets only make the parser to return list command, regardless of the user
option.