Update sort and find command to sort and search across new fields
Currently, the sort and find command are only able to function
with previous candidate fields.

The user is therefore not able to search across seniority,
interview status and application status.

Let's add the switch-case statements for the fields above,
and also the relevant ContainsKeywordsPredicate classes for the
find command.

Note: This commit also fixes the implementation for
updateSortedCandidateList() in the ModelManager. Error in the
previous implementation results in the original addressbook
candidate list being modified.