Enhance view command to specify time period
Currently, the view command only allows the user to view
all scheduled interviews.

The user is not able to filter the interview schedule by
time period.

Let's enhance the view command with three time period
keywords: 'today', 'week' and 'month', to display
scheduled interviews within the time period from when the
command is entered. Let's also modify the implementation
of InterviewSchedule related methods in the ModelManager
so that the interviews will be sorted from the earliest
to latest interview slot in the UI panel.