Add test cases for classes that relate to execution of view command
Currently, test cases relevant to the existing view command are out
of date due to modification to the view command accepted
user input format.

This causes an error when the test cases compile. No test cases are
present for the new ViewCommandParser class and Predicate classes.

Let's create more test cases to test the functionality of view
command to ensure it ouputs results we expect. In summary, this
includes:
* updates to test cases in ViewCommandTest class
* new ViewCommandParserTest class
* new WithinTimePeriodPredicate test classes
Due to the time sensitive input of view command, the ViewCommandTest
class is only able to check for the new ViewCommand object,
but is not able to handle an equality check.

Note: Slight implementation change to predicate subcless for
WithinTimePeriodPredicate to check that the interview date time
is later than the current date time.