feat: Add undo and redo feature to HackNet
When a user makes an error(e.g. deleting a wrong person from HackNet),
the user has to manually undo the previous operation, adding the person
back into HackNet in this specific example. This may be too troublesome
for them and a undo feature would be very helpful in reverting HackNet
back to that state.

Undo and Redo function only available for Add and Remove commands,
more commands will be supported later.

Tests and documentation will be added in a later commit.

See also: #72