Add EditCompanyCommand
The app did not have an edit company command. There was also some
unnecessary repetition in the code between each of the edit commands.

To fix this, a new EditCompanyCommand is added. The parseTags()
function was also taken out of the code of each of the edit command
parsers and instead stored in ParserUtil to facilitate code reuse.
Similarly, the MESSAGE_NOT_EDITED is instead stored in
commons.core.Messages since it is used by all 3 edit commands.

* add the EditEventCommand
* move parseTags() to ParserUtil
* move MESSAGE_NOT_EDITED to commons.core.messages

Note: The code to edit all instances of companyNames in events/persons
whose companyName matches the name of the company before it was edited
has not been implemented.