Fix bugs present in the feature navigate user input history
When users navigate to a previous user input before clearing the commandline and re-entering a new command, the user input history order was messed up.
This was due to the indexing of the user input history not being reset each time the user enters an input.
Refactored the class from TextFieldStorage.java to UserInputHistory.java.