Add remark command for candidates
Currently, there is no command or functionality
for the user to record a remark for each candidate.

This means that the user is not able to record certain
points relating to each candidate.

Let's extend the command class to create a RemarkCommand.
The implementation follows similarly to the tutorial
for adding a command. The remark command is implemented as a
standalone command, i.e. any replacement of a candidate's
remark has to be done via the remark command user input.