Allow team name and skill names to have spaces inside
Arguments for teams and skills are separated by space in batchedit
command, which means that the team and skill name itself cannot include
a space.

Some skills and team names are better expressed with spaces inside.

Regex for splitting the keywords were changed from a space character to
a comma which or with or without whitespace characters around. Regex
that checks for valid team and skillname was also changed to allow for
spaces inside.