Add statistic value to the charts so that users can easily see the number.
This addition aims to make reading the pie charts easier. A bar chart is
used to replace a pie chart when viewing by faculty. This makes more
sense as we are keen to compare each faculty within the hall.

Now summarise command will not add values that are 0 to piechartwindow
so that the charts will be more minimal and more to the point.