Allow editing of teams and skills of a person to be cumulative
Using Edit command to edit a person's teams or skills is not cumulative.

Users would typically are better off with cumulative edit, as one would
normally find out that a person in the addressbook are capable of one
more skill that the user did not know of. In that case, user would not
want to type every skill the person has that the User already stored in
HackNet again.