Update storage for interview to include composition
As we need to cascade the update of applicant and position to interview,
there is a need to create composition when loading from json file

Right now, interview has include the applicant and position reference.
But this is base on isSameX method not equals. If in the future need to
change then can change