Fix bug with Import
When the user imports a corrupted JSON file and closes Trackermon,
all existing data would be removed. The correct action would be for
Trackermon to retain the existing data.

This can be achieved by storing Model's show list, which has yet to
be updated. This allows us to overwrite the corrupted JSON file with
the old copy of Trackermon's data.