Update uniqueness of person
Person used to be uniquely
indentified by his name.

This is not a good representation
of the real world. As two people
can have the same name.

Update the person to be indentified
by phone number and/or email.

Update all associated person
display to come with email.

Let's,
* Change the uniqueness of person
to be identified by email and/or phone
* Update the display to show person
with email
* Update test cases and methods