Edit Bar chart in summarise command to get rid of any now integer intervals
This is done by adding a method to get the highest number of covid
students across faculties. This is used to set the limit of the bar
chart.

The bar chart is initialised to go by interval of an integer 1, up to
the highest limit which is from the getter method as mentioned in the
paragraph above.