Fix `editlog` bug
Previously, `editlog` did not correctly handle cases where a user edited
a log to be the same as an existing log, because the wrong equality
check method was called (duplicate checks were being done by logs having
same title AND description, which is not allowed).

This bug has been fixed and one test case has been added to verify the
behaviour.