Modify MainWindow to dynamically reload insights
At show insights, previously no call was made to refresh the list when
the command was called. Now, the MainWindow component, when called to
switch tabs, will dynamically request for a new list from the logic
component.

As an aside, the proper way to do this is probably to have a
`UniqueInsightsList`, or alternatively to have a new field in
`CommandResult`, but both seem unnecessary given the time constraints.