Fix EditCommand allowing duplicate pets
Presently, equality check for pets within `Model` takes into account the
diet as well as appointment fields of pets. This results in pets with
exact same details but different diets and/or appointments being able to
be added via the edit command.

Let's remove the check for diet and appointment fields, such that such
duplicate pet cases will no longer be allowed.