Inform users of detection of partially invalid indices for edit command
Partially invalid indices for batch edit using edit command does not
inform user of the detection. Instead, it only results in execution of
modification for at least the valid indices as an attempt to try best
to edit the persons.

Not informing users of the partial failure results in user being unaware
of the incomplete status of the edit command. This negatively affects
the user's workflow.