Improve unit tests for AddCommandParser
The dummy variable names and values for userInput is not representative
of intended use. For example, team field of a person is meant to be
filled in with a name of a project team, and the value 'friend' is
not intuitive to show that.

Test cases do not cover the more recently added functionality of
AddCommandParsers such as to separate team and skill values by a comma
and whitespace(s) before and after it.