Update developer guide and PPP
Currently, the developer guide lacks details for the remark,
view, sort and find command in the manual testing and use case
sections.

The user is not able to view the basic test cases to execute
to ensure the app is behaving properly.

Let's add in the use cases and manual test cases to address
some basic and integrated functionality that should work
properly for the above commands in the developer guide.

Note: this commit also updates details for the PPP