Refactor FindCommand, FindCommandParser and UG for Find
FindCommand now throws ParseException when the input has multiple prefixes. For example, "find /n /p". This is due to the bugs found when multiple prefixes are used "find /n /p" will only allow "find /n" to be accessed but not "find /p" due to the code structure. However, this issue can be resolved by only allowing users to use 1 prefix.