Fix the bug that valid team name incurs error
Current ArgumentTokenizer::tokenize only recognizes a prefix in a string
if it is preceded by at least one space. The prefix for `edit` command's
option is `-`. Team name allows all alphanumeric characters and space,
and some special characters including `#`, `-`, and `+`. Having a team
name of `Project-Team` is successful when editing. However the name
`Project -Team` results in HackNet interpreting the `Team` as a option
parameter after the `-` prefix.

Changing the prefix is the simplest viable solution. It also improves
user experience because now option prefix `o/` has the same syntax as
the other prefixes, such as `n/`, establishing a consistent syntax.